correct RateLimitApp name and update kube files
