+++ kubernetes/rate-limit/Chart.yaml
@@ -1,0 +1,5 @@
+apiVersion: v1
+appVersion: "2.7.2"
+description: HMDA Rate Limit gRPC API
+name: rate-limit
+version: 2.7.2
