+++ kubernetes/rate-limit/templates/service.yaml
@@ -1,0 +1,46 @@
+apiVersion: v1
+kind: Service
+metadata:
+  labels:
+    service: {{ .Values.ambassador.name }}
+  name: {{ .Values.ambassador.name }}
+  annotations:
+    getambassador.io/config: |
+      ---
+      apiVersion: ambassador/v0
+      kind: Mapping
+      name: check_digit_mapping
+      prefix: /v2/public/uli/
+      rewrite: /uli/
+      service: {{ template "rate-limit.fullname" .}}:{{ .Values.grpc.port }}
+
+spec:
+  type: {{ .Values.ambassador.service.type }}
+  ports:
+  - name: {{ .Values.ambassador.name }}
+    port: {{ .Values.ambassador.port }}
+    targetPort: {{ .Values.ambassador.port }}
+  selector:
+    service: {{ .Values.ambassador.name }}
+
+
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: {{ template "rate-limit.fullname" . }}-grpc
+  labels:
+    app: {{ template "rate-limit.fullname" . }}
+    chart: {{ template "rate-limit.chart" . }}
+    release: {{ .Release.Name }}
+    heritage: {{ .Release.Service }}
+spec:
+  type: {{ .Values.grpc.type }}
+  ports:
+    - port: {{ .Values.grpc.port }}
+      targetPort: {{ .Values.grpc.targetPort }}
+      name: {{ .Values.grpc.name }}
+  selector:
+    app: {{ template "rate-limit.name" . }}
+    release: {{ .Release.Name }}
+
