+++ kubernetes/rate-limit/values.yaml
@@ -1,0 +1,59 @@
+# Default values for check-digit.
+# This is a YAML-formatted file.
+# Declare variables to be passed into your templates.
+
+replicaCount: 1
+
+image:
+  repository: hmda/rate-limit
+  tag: latest
+  pullPolicy: IfNotPresent
+
+grpc:
+  type: ClusterIP
+  port: 60080
+  targetPort: 60080
+  name: rate-limit
+
+ambassador:
+  name: ambassador-rate-limit
+  port: 80
+  service:
+    type: NodePort
+
+ingress:
+  enabled: false
+  annotations: {}
+    # kubernetes.io/ingress.class: nginx
+    # kubernetes.io/tls-acme: "true"
+  ContainerPort: 60080
+  hosts:
+    - rate-limit.local
+  tls: []
+  #  - secretName: chart-example-tls
+  #    hosts:
+  #      - chart-example.local
+
+resources:
+  limits:
+    memory: "1000Mi"
+    cpu: "1"
+  requests:
+    memory: "1000Mi"
+    cpu: "1"
+  # We usually recommend not to specify default resources and to leave this as a conscious
+  # choice for the user. This also increases chances charts run on environments with little
+  # resources, such as Minikube. If you do want to specify resources, uncomment the following
+  # lines, adjust them as necessary, and remove the curly braces after 'resources:'.
+  # limits:
+  #  cpu: 100m
+  #  memory: 128Mi
+  # requests:
+  #  cpu: 100m
+  #  memory: 128Mi
+
+nodeSelector: {}
+
+tolerations: []
+
+affinity: {}
